+++ src/AiBiet.Tools.Commit/CommitTool.cs
@@ -1,4 +1,5 @@
 using AiBiet.Core.Interfaces;
+using AiBiet.Core.Utilities;
 
 using Spectre.Console;
 using Spectre.Console.Cli;